+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Bump AGP to v7.1.0
 - Bump desugar JDK library to v1.1.5
 - Migrate `RecentPatientsScreen` to use view effects
+- Fix `OnboardingScreen` not opening registration screen
 
 ## On Demo